+++ app/utils.js
@@ -9,3 +9,12 @@ if (typeof File === "function") {
 		});
 	};
 }
+
+
+
+export const downloadUrl = (url, filename) => {
+	const a = document.createElement("a");
+	a.setAttribute("download", filename);
+	a.href = url;
+	a.click();
+};