+++ app/index.vue
@@ -1,11 +1,23 @@
 <template>
 	<body>
-		<Playground />
+		<keep-alive>
+			<component v-if="view" :is="view"></component>
+		</keep-alive>
 	</body>
 </template>
 
 <script>
-	import Playground from "./views/playground.vue";
+	import Vue from "vue";
+
+
+
+	const components = [
+		"playground",
+		"profiler",
+	];
+
+
+	components.forEach(name => Vue.component(name, () => import(`./views/${name}.vue`)));
 
 
 
@@ -13,8 +25,23 @@
 		name: "lotus",
 
 
-		components: {
-			Playground,
+		data () {
+			return {
+				view: null,
+			};
+		},
+
+
+		mounted () {
+			this.onHashChange ();
+			window.onhashchange = () => this.onHashChange();
+		},
+
+
+		methods: {
+			onHashChange () {
+				this.view = location.hash.substr(1) || "playground";
+			},
 		},
 	};
 </script>