+++ app/views/playground.vue
@@ -517,6 +517,26 @@
 					width: 100%;
 				}
 
+				// sheet custom styles
+				.sheet.live
+				{
+					white-space: nowrap;
+					display: inline-block;
+
+					.page
+					{
+						display: inline-block;
+						margin: 1em;
+						background: #f6fffa;
+						border-radius: 1em;
+					}
+
+					.cursor
+					{
+						fill: lightblue;
+					}
+				}
+
 				&.inspecting
 				{
 					.token.mismatched use