+++ index.js
@@ -1,4 +1,0 @@
-
-// things to export are not ready yet.
-
-// TODO: create the file of index.d.ts