+++ index.ts
@@ -1,0 +1,8 @@
+
+import * as staffSvg from "./inc/staffSvg";
+
+
+
+export {
+	staffSvg,
+};